::Logging FFI
