+++ 3DRadSpace/Engine3DRadSpace/FFI/Logging/Message.h
