+++ 3DRadSpace/Engine3DRadSpace/FFI/Logging/Warning.cpp
@@ -1,0 +1,14 @@
+#include "Warning.h"
+#include "../../Logging/Warning.hpp"
+
+using namespace Engine3DRadSpace::Logging;
+
+void E3DRSP_SetLastWarning(const E3DRSP_Warning* warning)
+{
+	SetLastWarning(Warning(
+		warning->Code,
+		warning->Details,
+		warning->Severity,
+		warning->Extra
+	));
+}
