+++ 3DRadSpace/Engine3DRadSpace/Logging/Message.cpp
@@ -1,5 +1,4 @@
 #include "Message.hpp"
-#include "Message.h"
 #include <fstream>
 #include <print>
 #include <iostream>
@@ -35,13 +34,3 @@ inline void Engine3DRadSpace::Logging::DefaultMessageHandler(const Message &msg)
 }
 
 inline FuncMessageHandler Engine3DRadSpace::Logging::MessageHandler = DefaultMessageHandler;
-
-void E3DRSP_SetLastMessage(const E3DRSP_Message* message)
-{
-	SetLastMessage(Message(
-		message->Code,
-		message->Details,
-		message->Extra
-		)
-	);
-}
